Add solution for 'Sort Items by Groups Respecting Dependencies' problem
Implemented a Golang solution for 'Sort Items by Groups Respecting Dependencies', problem 1203. The implementation involves applying topological sort twice: once for groups and once for items within each group. The solution also handles special cases including but not limited to items that belong to no group, multiple viable solutions, and conditions where a solution can't be found. The main motive of the solution is to return a sorted list of items such that items belonging to the same group are adjacent and every item comes after those in its `beforeItems` list.