Add solution for 'Unique Paths' problem in Golang
Implemented a Golang solution for 'Unique Paths', problem 62 in the Medium category. The task is to calculate possible unique paths a robot can make to reach the lower-right position of a n x m 2D grid considering the robot only can move down or to the right. This is achieved by creating a dp grid of size 'n', iterating over 'm' and 'n', and then updating the dp grid based on the upper cell and the cell on the left.