Add solution for splitting linked list into parts
Implemented the function `splitListToParts` in Golang to solve the problem of splitting a singly linked list into k parts as equally as possible. This solution is part of the Medium level problems in Golang. Exploring the linked list to calculate the overall length, the function subsequently calculates the width and remainder. It then iteratively splits the linked list into the desired parts and assigns them into an array. The parts are in order of occurrence in the original linked list.