Add Combination Sum IV problem solution
The new Python file 377_CombinationSumIV.py was added under Medium folder. The file implements a function `combinationSum4` which calculates the number of possible combinations that add up to a specific target number, using a given list of numbers.

This function uses dynamic programming to solve the problem efficiently. The if statement in the for loop is a boundary condition that ensures the function does not attempt to access a nonexistent index of the results array. A test function is also included to verify the correctness of the solution.