The provided commit message does not match the diff. Here is the correct commit message:
Commit message:

Add solution for 377 Combination Sum IV problem

Added a solution for the Medium level Golang problem "377 Combination Sum IV". This solution implements a dynamic programming approach to calculate the total combinations of numbers from an array that can sum up to a target value. The solution defines a function `combinationSum4` that takes an array of integers and a target integer as input and provides the number of combinations as output. The created 'dp' array represents the total combinations for each sum up to the target.