The provided commit message seems to be incorrectly describing the changes. You need to submit your response to generate a correct commit message. Here is the revised version:
Commit message:

Added solution for "1658 Minimum Operations to Reduce X to Zero"

This commit adds the solution to the LeetCode problem "1658. Minimum Operations to Reduce X to Zero". The added function, minOperations, finds the optimal solution to remove elements from an array from either end and subtract its value to reduce the given number x to zero. If not possible, it returns -1. The function employs a map to keep track of the previously visited values, thus optimizing the time complexity. The solution has been tested with provided examples.