Add solution for 'Is Subsequence' problem in Golang
This commit adds a new file which provides a solution for 'Is Subsequence' problem of leetcode written in Go language. The solution determines whether one string is a subsequence of another and returns a boolean value accordingly. The solution handles large incoming sequences effectively by iterating the larger string until all the characters of the smaller string have been found.