Implemented solution for 'Remove Colored Pieces' problem
Added solution in Golang for the '2038_Remove Colored Pieces if Both Neighbors are the Same Color' problem. The game rules were implemented as conditions in a for loop iterating through the colors string. Counter variables for 'A' and 'B' colors were compared at the end to determine the winner. The function has been tested with the provided problem examples.