Your commit message doesn't match the actual change. Here is the appropriate one:
Add solution for Number of Flowers in Full Bloom problem

Added a new Golang solution for a problem: "Number of Flowers in Full Bloom". The function fullBloomFlowers calculates the number of flowers that are in full bloom at the moment each person arrives to see the flowers. It sorts the start and end times of the flowers' bloom period and searches for the arrival time of each person using built-in Go functions. This is categorized as a hard level problem.