Add solution for 1425 Constrained Subsequence Sum problem
New Golang solution added for the Hard level problem "1425 Constrained Subsequence Sum". This dynamic programming-based solution calculates the maximum sum of a non-empty array subsequence with given constraints. It generates a dynamic array 'dp' to store the sum sequences, 'q' slice to maintain sequence indices and 'ans' to hold the maximum sum. Function `constrainedSubsetSum` is established that takes in the array and constraint parameter.