Add solution for Poor Pigs problem
Added golang solution for the hard-level problem 'Poor Pigs' on LeetCode. This solution uses the concept of base conversion to determine the minimum number of pigs needed to find the poisonous bucket within the given time constraints. Moreover, the function to solve this problem, poorPigs, has been tested on several test cases in the main function, thus validating its correctness.