Add solution for 'Find The Original Array' problem
Added golang solution for the medium-level problem 'Find The Original Array of Prefix Xor' on LeetCode. This program finds and returns an array that satisfies the given condition with the help of bitwise-xor operation. It includes a main function where this solution is validated by using an example test case.