Add solution for 'Determine if a Cell Is Reachable at a Given Time' problem
Implemented a function 'isReachableAtTime' in Golang to solve the given problem in LeetCode. It determines if a cell on an infinite 2D grid is reachable from a starting cell in a given amount of time. Both start and end points are parameters, along with the time taken. Utilizes basic arithmetic functions to calculate if the destination can be reached. It also includes a main function that demonstrates the usage of this function on example inputs.