Add solution for 'Find Unique Binary String' problem
Implemented a function in Golang to solve the 'Find Unique Binary String' problem from Medium category. The solution uses bitwise operations and strings manipulation to obtain a binary string of length n that is not present in the input array. The main function contains examples of how to use this function.