Add solution for 'Reduction Operations to Make the Array Elements Equal' problem
Implemented a function to solve the 'Reduction Operations to Make the Array Elements Equal' problem in Golang. The function counts the number of operations needed to make all elements in an integer array equal. The function implements sort and for loop to achieve this. Test cases are included in the main function.