Add solution for 'Maximum Number of Coins You Can Get' problem
Implemented a function to solve the 'Maximum Number of Coins You Can Get' problem in Golang. With this function, based on the provided number of piles and their respective amounts of coins, the user can calculate the maximum number of coins they can have after choosing any 3 piles of coins in each step. The function employs the "sort" library for preparing the piles in ascending order and then applies a for loop to calculate the final sum of coins. Test cases are provided within the main function.